JohnCMS 4.0.0
[new]   Новое, быстродействующее ядро системы
[new]   Мультиязычность, возможность загружать / удалять языки
[new]   Новая Админ Панель
[new]   Определение мобильного браузера (глобальная переменная с кэшированием)
[new]   Новый модуль комментариев (глобальный механизм)
        - возможность для пользователей управлять комментариями
        - очистка сообщений выбранного пользователя
        - ответ на комментарий
[new]   Карта сайта
        - оптимизация для SAPE, LinkFeed, разбивка на блоки по 150 ссылок
        - поднятие всех тем форума и статей в библиотеке на 3-й уровень вложенности
        - применено кэширование, для ускорения работы и снижения нагрузки
        - определение мобильных браузеров (им карта не показывается)
[new]   Личные фотоальбомы пользователей
        - создание, удаление, редактирование, сортировка альбомов
        - назначение прав доступа к альбомам (всем, друзьям, пароль, закрыто)
        - редактирование фотографий (яркость, контрастность, поворот)
        - редактирование описаний, перемещение между альбомами
        - рейтинги и многое другое...
[new]   Личные Гостевые (стена) пользователей
        - пользователь сам управляет своей Гостевой
        - уведомление о новых комментариях
[new]   Новая Анкета пользователя
        - расширенная информация
        - статистика активности
[new]   Новый модуль Кармы, алгоритм прежний, но переписан код
[new]   Новый Рекламный модуль, возможность ставить "Прямые" ссылки
[new]   Предпросмотр картинок файл-обменника Форума. Оптимизация кода
[new]   Новый модуль модерации регистраций пользователей
        - массовое подтверждение / удаление регистраций
        - удаление всех регистраций с выбранным адресом IP
[new]   Фиксируется полная история IP адресов для каждого посетителя
        - просмотр истории IP посетителя в хронологическом порядке
        - поиск (для Администрации) по всей истории IP
[new]   Подсветка поисковых запросов форума (Krite + AlkatraZ)
[new]   Цветные графики голосований (от VARG)
[new]   Панель быстрого ввода BBcode (от seg0ro)
[new]   Новый модуль F.A.Q. со встроенными каталогами Аватаров и Смайлов
[new]   Полностью переработана функция "история нарушений"
[new]   Общее хранилище /files для выгружаемых файлов всех модулей
[new]   Управление доступом к просмотру Актива сайта для Гостей
[mod]   При закрытии модуля, на него скрываются ссылки с Главной
[fix]   В закрытой теме форума, оставались открытыми голосования
[del]   Из дистрибутива удален Чат

+++ download/dfile.php
@@ -15,12 +15,11 @@
 */
 
 defined('_IN_JOHNCMS') or die('Error: restricted access');
-
-require_once ("../incfiles/head.php");
+require_once("../incfiles/head.php");
 if ($rights == 4 || $rights >= 6) {
     if ($_GET['file'] == "") {
-        echo "Не выбран файл<br/><a href='?'>К категориям</a><br/>";
-        require_once ('../incfiles/end.php');
+        echo $lng_dl['file_not_selected'] . "<br/><a href='?'>" . $lng['back'] . "</a><br/>";
+        require_once('../incfiles/end.php');
         exit;
     }
     $file = intval(trim($_GET['file']));
@@ -28,19 +27,22 @@ if ($rights == 4 || $rights >= 6) {
     $file2 = mysql_num_rows($file1);
     $adrfile = mysql_fetch_array($file1);
     if (($file1 == 0) || (!is_file("$adrfile[adres]/$adrfile[name]"))) {
-        echo "Ошибка при выборе файла<br/><a href='?'>К категориям</a><br/>";
-        require_once ('../incfiles/end.php');
+        echo $lng_dl['file_not_selected'] . "<br/><a href='?'>" . $lng['back'] . "</a><br/>";
+        require_once('../incfiles/end.php');
         exit;
     }
-    $refd = mysql_query("select * from `download` where type = 'cat' and id = '" . $adrfile[refid] . "';");
+    $refd = mysql_query("select * from `download` where type = 'cat' and id = '" . $adrfile[refid] . "'");
     $refd1 = mysql_fetch_array($refd);
-    unlink("$adrfile[adres]/$adrfile[name]");
-    mysql_query("delete from `download` where id='" . $adrfile[id] . "' LIMIT 1;");
-    echo "Файл удалён<br/>";
+    if (isset($_POST['submit'])) {
+        unlink("$adrfile[adres]/$adrfile[name]");
+        mysql_query("delete from `download` where id='" . $adrfile[id] . "' LIMIT 1;");
+        echo '<p>' . $lng_dl['file_deleted'] . '</p>';
+    } else {
+        echo '<p>' . $lng['delete_confirmation'] . '</p>' .
+            '<form action="index.php?act=dfile&amp;file=' . $file . '" method="post">' .
+            '<input type="submit" name="submit" value="' . $lng['delete'] . '" />' .
+            '</form><p><a href="index.php?act=view&amp;file=' . $file . '">' . $lng['cancel'] . '</a></p>';
+    }
 }
-else {
-    echo "Нет доступа!";
-}
-echo "&#187;<a href='?cat=" . $refd1[id] . "'>В папку</a><br/>";
-
+echo "<p><a href='?cat=" . $refd1[id] . "'>" . $lng['back'] . "</a></p>";
 ?>
